Fix BasicBrowsing's TR links. Consolidate TrackingProtectionTests
